requirements.txt
pydantic==2.11.7
pydantic-ai==0.2.9
mcp==1.9.1
SQLAlchemy==2.0.41
langchain-mcp-adapters==0.1.1
langgraph==0.6.6
langchain==0.3.25
langchain-openai==0.3.31